+++ CHANGELOG.md
@@ -12,6 +12,11 @@ ENHANCEMENTS
 
 - Move AWS IAM Identity Center permission set resources to a sub-module. ([#150](https://github.com/schubergphilis/terraform-aws-mcaf-landing-zone/pull/150))
 
+<<<<<<< HEAD
+=======
+- Allow principal exceptions on `aws_deny_disabling_security_hub` and `aws_deny_leaving_org` organizations policy. Move SCP's variables into `aws_service_control_policies` ([#153](https://github.com/schubergphilis/terraform-aws-mcaf-landing-zone/pull/153)).
+  
+>>>>>>> f7a6cca (adding updated to README.md and UPGRADING.md)
 ## 0.19.1 (2022-12-14)
 
 ENHANCEMENTS